+++ packages/graphql-server-micro/src/microApollo.ts
@@ -1,6 +1,6 @@
 import { GraphQLOptions, HttpQueryError, runHttpQuery } from 'graphql-server-core';
 import * as GraphiQL from 'graphql-server-module-graphiql';
-import { json } from 'micro';
+import { createError, json, RequestHandler } from 'micro';
 import * as url from 'url';
 import {IncomingMessage, ServerResponse} from 'http';
 
@@ -8,7 +8,7 @@ export interface MicroGraphQLOptionsFunction {
   (req?: IncomingMessage): GraphQLOptions | Promise<GraphQLOptions>;
 }
 
-export function microGraphql(options: GraphQLOptions | MicroGraphQLOptionsFunction) {
+export function microGraphql(options: GraphQLOptions | MicroGraphQLOptionsFunction): RequestHandler {
   if (!options) {
     throw new Error('Apollo Server requires options.');
   }
@@ -29,47 +29,48 @@ export function microGraphql(options: GraphQLOptions | MicroGraphQLOptionsFuncti
       query = url.parse(req.url, true).query;
     }
 
-    runHttpQuery([req, res], {
-      method: req.method,
-      options: options,
-      query: query,
-    }).then((gqlResponse) => {
+    try {
+      const gqlResponse = await runHttpQuery([req, res], {
+        method: req.method,
+        options: options,
+        query: query,
+      });
+
       res.setHeader('Content-Type', 'application/json');
-      res.write(gqlResponse);
-      res.end();
-    }, (error: HttpQueryError) => {
-      if ( 'HttpQueryError' !== error.name ) {
-        throw error;
+      return gqlResponse;
+    } catch (error) {
+      if ('HttpQueryError' === error.name) {
+        if (error.headers) {
+          Object.keys(error.headers).forEach((header) => {
+            res.setHeader(header, error.headers[header]);
+          });
+        }
       }
 
-      if ( error.headers ) {
-        Object.keys(error.headers).forEach((header) => {
-          res.setHeader(header, error.headers[header]);
-        });
+      if (!error.statusCode) {
+        error.statusCode = 500;
       }
 
-      res.statusCode = error.statusCode;
-      res.write(error.message);
-      res.end();
-    });
+      throw error;
+    }
   };
 }
 
-export function microGraphiql(options: GraphiQL.GraphiQLData) {
-  return (req: IncomingMessage, res: ServerResponse) => {
-    const q = req.url && url.parse(req.url, true).query || {};
-    const query = q.query || '';
-    const operationName = q.operationName || '';
+export interface MicroGraphiQLOptionsFunction {
+  (req?: IncomingMessage): GraphiQL.GraphiQLData | Promise<GraphiQL.GraphiQLData>;
+}
 
-    const graphiQLString = GraphiQL.renderGraphiQL({
-      endpointURL: options.endpointURL,
-      query: query || options.query,
-      variables: q.variables && JSON.parse(q.variables) || options.variables,
-      operationName: operationName || options.operationName,
-      passHeader: options.passHeader,
+export function microGraphiql(options: GraphiQL.GraphiQLData | MicroGraphiQLOptionsFunction): RequestHandler {
+  return (req: IncomingMessage, res: ServerResponse) => {
+    const query = req.url && url.parse(req.url, true).query || {};
+    GraphiQL.resolveGraphiQLString(query, options, req).then(graphiqlString => {
+      res.setHeader('Content-Type', 'text/html');
+      res.write(graphiqlString);
+      res.end();
+    }, error => {
+      res.statusCode = 500;
+      res.write(error.message);
+      res.end();
     });
-    res.setHeader('Content-Type', 'text/html');
-    res.write(graphiQLString);
-    res.end();
   };
 }